sophia_iri

Module resolve

Implementation of IRI resolution as per [RFC 3987].

This module is based on https://docs.rs/oxiri/.

NB: compared to Iri and IriRef, BaseIri and BaseIriRef are slower to build, because they analyse the internal structure of the IRI, in order to allow for efficient resolution of relative IRIs.

Structs

BaseIri

A BaseIri is an absolute IRI against which relative IRIs can be resolved. It stores the internal structure of the IRI, to allow for efficient resolution of relative IRIs against itself.

BaseIriRef

A BaseIriRef is an absolute or relative IRI reference, against which relative IRIs can be resolved. It stores the internal structure of the IRI, to allow for efficient resolution of relative IRIs against itself.

IriParseError

An error raised during Iri or IriRef validation.

Oxiri

A RFC 3987 IRI.

OxiriRef

A RFC 3987 IRI reference.

Traits

Resolvable

A trait for anything that can be resolved against a BaseIri or a BaseIriRef.