sophia_iri::resolve

Trait Resolvable

pub trait Resolvable<T: Borrow<str>>: Borrow<str> {
    type OutputAbs;
    type OutputRel;

    // Required methods
    fn output_abs(res: Result<T, IriParseError>) -> Self::OutputAbs;
    fn output_rel(res: Result<T, IriParseError>) -> Self::OutputRel;
}

A trait for anything that can be resolved against a BaseIri or a BaseIriRef.

Required Associated Types

type OutputAbs

The output type when joining to an absolute base.

type OutputRel

The output type when joining to an relative base.

Required Methods

fn output_abs(res: Result<T, IriParseError>) -> Self::OutputAbs

Method for producing the Self::OutputAbs from a raw result.

fn output_rel(res: Result<T, IriParseError>) -> Self::OutputRel

Method for producing the Self::OutputRel from a raw result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.